requirements.txt
fastapi[standard]
pydantic
openai
python-dotenv